+++ examples/heterogeneous.py
@@ -41,7 +41,7 @@ model = ResSim(Lx=1, Ly=1, Nx=8, Ny=8,
                inj_xy=[[0, 0]], inj_rates=[[1]],
                prd_xy=[[1, 1]], prd_rates=[[1]])
 
-model.assemble_wells(None, 0)
+model.assemble_wells(None, None, 0)
 [P_coarse, V] = model.TPFA(model.K)
 
 ax = axs[0, 0]
@@ -64,7 +64,7 @@ ax.set(title="log-Permeability", aspect="equal")
 cc = ax.pcolormesh(logK.T[..., 0], edgecolors='k', linewidth=.01, cmap="jet")
 fig.colorbar(cc, axs[1, 1], orientation="horizontal")
 
-model.assemble_wells(None, 0)
+model.assemble_wells(None, None, 0)
 [P_fine, V] = model.TPFA(model.K)
 
 ax = axs[0, 2]